requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
PyMySQL==1.1.0

# Session Management
Flask-Session==0.5.0

# Authentication
PyJWT==2.8.0

# Environment Variables
python-dotenv==1.0.0

# Data Processing
pandas==2.1.0
openpyxl==3.1.2
python-dateutil==2.8.2

# AI Integration
openai==1.3.0

# Report Generation
reportlab==4.0.9
xlsxwriter==3.1.9
Pillow==10.1.0

# PDF Generation (한글 지원)
pypdf2==3.0.1

# HTTP Server
gunicorn==21.2.0

# Utils
Werkzeug==3.0.0

# File Type Detection (보안)
python-magic-bin==0.4.14; sys_platform == 'win32'
python-magic==0.4.27; sys_platform != 'win32'
